curriculum/shared/requirements.txt
# LLM providers
anthropic>=0.40.0
openai>=1.50.0
google-generativeai>=0.8.0

# Embeddings + vector stores
chromadb>=0.5.0
sentence-transformers>=3.0.0

# RAG utilities
tiktoken>=0.7.0
langchain-text-splitters>=0.3.0

# HTTP + async
httpx>=0.27.0
aiohttp>=3.10.0

# Data
python-dotenv>=1.0.0
pydantic>=2.8.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0

# Observability (optional)
langfuse>=2.0.0

# Utilities
rich>=13.0.0
tenacity>=9.0.0
